[feat] project: refactor the struct to delete public folder
