--- public/src/modules/redux-thunk/thunk.js
@@ -1,5 +1,0 @@
-import { createThunkMiddleware } from './createThunkMiddleware'
-
-const thunk = createThunkMiddleware()
-
-export default thunk
