--- public/src/redux/reducers/index.js
@@ -1,3 +1,0 @@
-import { combineReducers } from '../../modules/redux'
-
-export const reducer = combineReducers({})
