--- public/src/redux/store.js
@@ -1,8 +1,0 @@
-import { applyMiddleware, createStore } from '../modules/redux'
-import { middlewares } from './middlewares/index.js'
-import { reducer } from './reducers/index.js'
-
-const initialState = {}
-
-const createStoreWithMiddleware = applyMiddleware(middlewares)(createStore)
-export const store = createStoreWithMiddleware(reducer, initialState)
