+++ server/index.js
@@ -1,5 +1,4 @@
 /* eslint-disable no-console */
-
 import { execSync } from 'child_process'
 import express from 'express'
 import path from 'path'
@@ -10,36 +9,47 @@ const __filename = fileURLToPath(import.meta.url)
 const __dirname = path.dirname(__filename)
 
 const app = express()
+const PORT = 3000
 
+const clientPath = path.join(__dirname, '../src')
+const buildPath = path.join(__dirname, '../build')
+const handlebarsPath = path.join(__dirname, '../src/handlebars')
+
+// Middleware логгера
 app.use((req, res, next) => {
 	res.setHeader('X-Powered-By', 'Express')
 	console.log(`[${req.method}] ${req.url}`)
 	next()
 })
 
-// Обслуживаем статические файлы
-app.use('/src', express.static('src'))
-app.use('/reload', express.static('node_modules/reload'))
+// 📦 СТАТИКА
+app.use('/src', express.static(clientPath))
+app.use('/build', express.static(buildPath))
+app.use('/src/handlebars', express.static(handlebarsPath))
+app.use(
+	'/reload',
+	express.static(path.join(__dirname, '../node_modules/reload')),
+)
 
-// Только для не-статических путей отправляем index.html
-app.get('/', (_, response) => {
-	response.sendFile(path.join(__dirname, '../index.html'))
-})
-
-app.get('/*path', (req, response, next) => {
-	// Пропускаем статические файлы и reload
-	if (req.url.startsWith('/public/') || req.url.startsWith('/reload/')) {
+// 🔄 SPA fallback
+app.use((req, res, next) => {
+	if (
+		req.url.startsWith('/src/') ||
+		req.url.startsWith('/build/') ||
+		req.url.startsWith('/reload/')
+	) {
 		return next()
 	}
-	response.sendFile(path.join(__dirname, '../index.html'))
+	res.sendFile(path.join(clientPath, 'index.html'))
 })
 
-app.listen(3000, () => {
-	console.log('Server running on http://127.0.0.1:3000')
+// 🚀 Запуск сервера
+app.listen(PORT, () => {
+	console.log(`✅ Server running at http://127.0.0.1:${PORT}`)
 })
 
-reload(app).then(function () {
+// 🔁 Reload + build
+reload(app).then(() => {
 	execSync('npm run build', { stdio: 'inherit', shell: true })
 })
-
 /* eslint-enable no-console */
