+++ src/modules/redux-thunk/thunk/index.js
@@ -1,0 +1,5 @@
+import { createThunkMiddleware } from '../createThunkMiddleware'
+
+const thunk = createThunkMiddleware()
+
+export default thunk
